Added SWAL2

+++ src/app/clientes/eliminar-cliente/eliminar-cliente.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { ClientesService } from '../clientes.service';
 import { Cliente } from '../cliente.model';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-eliminar-cliente',
@@ -23,8 +24,32 @@ export class EliminarClienteComponent implements OnInit {
 
   eliminar() {
     if (this.id) {
-      this.clientesService.eliminarCliente(this.id);
-      this.id = null;
+      Swal.fire({
+        title: '¿Estás seguro?',
+        text: 'No podras revertir este cambio',
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonColor: '#3085d6',
+        cancelButtonColor: '#d33',
+        confirmButtonText: 'Sí, eliminar!',
+        cancelButtonText: 'Cancelar'
+      }).then((result) => {
+        if (result.value) {
+          this.clientesService.eliminarCliente(this.id);
+          this.id = null;
+          Swal.fire(
+            'Eliminado!',
+            'El registro se elimino con éxito.',
+            'success'
+          );
+        } else {
+          Swal.fire(
+            'No se elimino!',
+            'No hubieron cambios.',
+            'warning'
+          );
+        }
+      });
     }
   }
 }
